Removed some basic debug prints.
I have discovered that the issues surrounding the use of multiple
databases for CCOMData are related to the Asset entity. It seems that
Asset is ignoring any atempts I make at changing the database. It is
as though all information is cached away and cannot be changed without
restarting the app.